Notifications Done

+++ database/migrations/2024_11_16_191159_create_permissions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->string('header',255);
             $table->text('description')->nullable();
             $table->tinyInteger('status')->default(1);
-            $table->unsignedBigInteger('parent_id')->nullable();
+            $table->unsignedInteger('parent_id')->nullable();
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent()->useCurrentOnUpdate();
